refactor: update Synthetix account ID type to BigInt and fix deposit flow by removing redundant blockchain lookups

+++ lib/account/bloc/user_account_state.dart
@@ -96,7 +96,7 @@ class ReadyToTradeState extends UserAccountState {
   });
 
   final String walletAddress;
-  final int synthetixAccountId;
+  final BigInt synthetixAccountId;
 
   @override
   bool get isWalletConnected => true;
